requirements.txt
pandas
python-docx
requests
beautifulsoup4